Fix to hybrid BVH initialization
 * we must initialize nDaughters before passing it down to BuildStructure
 * on linux the sequence of initialization was not as intended (on clang-Mac it worked)

This fix is due to Mihaly Novak

+++ test/core/TestHybridBVH.cpp
@@ -46,7 +46,7 @@ void QueryStructure(BVHStructure const &s)
   };
 
   int c = 0;
-  for (int i = 0; i < s.fNumberOfOriginalBoxes; ++i) {
+  for (int i = 0; i < (int)s.fNumberOfOriginalBoxes; ++i) {
     c += (i + 1);
   }
